Help Center: Declare request test double
